protocol v2: emit ERR, don't die()
Change the recently added packet_client_error() helper to not die()
after emitting an ERR line, but rather exit(128). This ensures that we
get sensible error messages, as the client won't need to emit both the
ERR line, and the redundant death message from the server.

As tests added in preceding commits show doing this was racy,
sometimes we'd get ERR before die() from a remote, or the other way
around.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>