SANITIZE tests: fix memory leaks in t0500*, add to whitelist
Fix a couple of memory leaks in running t0500* and remove the
SANITIZE_LEAK prerequisite on the one test we were skipping.

The trace2 issue is a memory leak introduced in ee4512ed48 (trace2:
create new combined trace facility, 2019-02-22), that was seemingly
done by omission. We already had a "unset_self" function to free the
resources, let's free the thread name there.

The other issue was a leak in the progress.c test helper itself, added
in 2bb74b53a4 (Test the progress display, 2019-09-16).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>