SANITIZE tests: fix leak in mailmap.c
Get closer to being able to run t4203-mailmap.sh by fixing a memory leak in mailmap.c.

In the free_mailmap_entry() code added in 0925ce4d49 (Add map_user()
and clear_mailmap() to mailmap, 2009-02-08) the intent was clearly to
do this, but while we freed parts of the mailmap_entry structure, we
didn't free the structure itself.