test-tool run-command: fix buggy init pattern
In be5d88e1128 (test-tool run-command: learn to run (parts of) the
testsuite, 2019-10-04) an init pattern was added that would use
TESTSUITE_INIT, but then promptly memset() everything back to 0. We'd
then set the "dup" on the two string lists.

Let's just use the init macro and stop doing the redundant memset,
since the "next" was set from -1 to 0 we can start it at 0 too.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>