transport-helper: add and use *_release() function
Plug a memory leak in the transport-helper code, we wouldn't free the
"data->name", and on disconnect of the main transport we wouldn't free
the helper at all.

One of the two callers of transport_helper_init() calls it with a
string that's owned by the "struct remote", and in other one we
returned a truncated xstrndup()'d string. Let's always pass in a
"const char *" and have the transport_helper_init() xstrdup() it,
which makes memory management easy.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>