transport.c: replace a xmemdupz() with xstrndup()
Change the xmemdupz() added in 6b33894f997 (transport: simplify
duplicating a substring in transport_get() using xmemdupz(),
2014-12-24) to a xstrndup().

This would be introducing a bug if it was important that we padded the
string to the "len" with extra NUL bytes, but that's not the case. See
the original code in 6eb996b5707 (Add support for external programs
for handling native fetches, 2009-08-05) which treats it as a C
string.

The reason we ended up with xmemdupz() here is due to code added in
5d5cc488a7 (Pass unknown protocols to external protocol handlers,
2009-12-09) using strncpy(), which does pad the string. Let's change
this to a xstrndup() instead to make it clear that that property of it
isn't important.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>