http: drop support for curl < 7.17.0
In the last commit we dropped support for curl < 7.16.0, let's
continue that and drop support for versions older than 7.17.0. This
allows us to simplify the code by getting rid of some "#ifdef"'s to do
with the CURLOPT_USE_SSL flag.

The CURLOPT_USE_SSL flag was added in 7.17.0 as an alias for
CURLOPT_FTP_SSL[1].

The documentation[1] currently claims that it was introduced in
7.16.4, but the symbols-in-versions file correctly states
7.17.0[2].

I've submitted an upstream
patch (<patch-1.1-953bab490-20210730T170510Z-avarab@gmail.com>) to the
curl-library mailing list fix the documentation.

1. https://curl.se/libcurl/c/CURLOPT_USE_SSL.html
2. https://github.com/curl/curl/blob/master/docs/libcurl/symbols-in-versions

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>