progress.c tests: support old-style usage
In a preceding commit I changed "test-tool progress" to only support
input on stdin, let's support the old-style mode we used to support
before, this is to help with a semantic conflict with other patches of
mine.

We'll eventually be able to revert this as part of a cleanup after the
in-flight patches have landed.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>