bundle-uri client: add boolean transfer.bundleURI setting
The yet-to-be introduced client support for bundle-uri will always
fall back on a full clone, but we'd still like to be able to ignore a
server's bundle-uri advertisement entirely. This is useful for
testing, and if a server is pointing to bad bundles, they take a while
to time out etc.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>