fetch-negotiator: call BUG() on API misuse, don't segfault
As noted in ec06283844a (fetch-pack: introduce negotiator API,
2018-06-14) it's important that the fetch negotiator's callbacks be
called in the documented order, and that some of them never be called
again after other "later" callbacks are called.

But let's assert that with a BUG(), instead of setting the relevant
callbacks to NULL. We'll now give a meaningful error on API misuse,
instead of segfaulting.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>