index-pack: make --fix-thin play nice with connectivity check
In c6807a40dcd (clone: open a shortcut for connectivity check,
2013-05-26) the index-pack command was made to exit non-zero if any
foreign objects were left over under
--check-self-contained-and-connected.

This doesn't make much sense, and precludes us from using
--check-self-contained-and-connected in combination with --fix-thin in
a way that'll signal useful errors on non-zero exit.

This is because in conclude_pack() (which we'll call under --fix-thin)
we'll already die() if we have unresolved deltas, i.e. unaccounted-for
objects. This happens under both --fix-thin and without, under
--fix-thin we'll attempt to fix up the pack using existing repository
objects.

is redundant to the checking our internal callers need from
index-pack (--check-self-contained-and-connected is "For internal use
only"), and

It's redundant because in conclude_pack()

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>