format-patch tests: simplify recently added tests with a helper
Add a new check_headers() helper and use it for the tests added in
db91988aa1 (format-patch: allow a non-integral version numbers,
2021-03-23). Just doing a test_cmp on the "[...]" part of the subject
line is both more legible, and makes for a more accurate test.

Currently we ignore any header except "Subject: ", but as the name of
the function suggests we'll extend this in subsequent commits for
sanity checking other basic headers, splitting the arguments across
lines is likewise done to make the diff of those subsequent commits
smaller.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>