format-patch tests: test existing subject option behavior
Test the behavior of the subject manipulation options. There were some
tests added for the -v option in in db91988aa1 (format-patch: allow a
non-integral version numbers, 2021-03-23), but not for the basic
functionality of the other ones. There was no test for --rfc at all,
nor for the basic usage of --subject-prefix.

I consider some of this behavior to be buggy, e.g. that we don't die
with usage() if both --subject-prefix and --rfc are provided (one will
clobber the other).

But let's test for the established behavior as-is for now. I have
another series that fixes that behavior (and more) that I'm planning
to submit at some point, but doing so is a much larger change than
adding these tests.

As noted in the preceding commit the check_headers() function will be
expanded soon to check other basic headers, so let's title the tests
accordingly.

I came up with the "negate" idiom here as an alternative to the
verbose alternative of duplicating the command between the "if" and
"else" branches, see 69576367921 (merge-ort: expand only for
out-of-cone conflicts, 2021-09-08) for one recent example of the more
verbose alternative.

But having done that I discovered an existing use in the test suite,
see cf0ad4d1999 (cherry-pick/revert: honour --no-gpg-sign in all case,
2020-04-03). So we can be assured that this is portable. It wouldn't
work for putting a "!" at the start of the line (so we couldn't use it
for the code added in 69576367921 (merge-ort: expand only for
out-of-cone conflicts, 2021-09-08)) , but for a "test_must_fail" it'll
work just fine.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>