parse-options.c: add and use PARSE_OPT_ERROR_AT_NON_OPTION flag
Add a new PARSE_OPT_ERROR_AT_NON_OPTION flag, with this flag we'll
error out with a usage message if there's any remaining items in argv,
not just if we find unknown options. Without this parse_options() will
ignore any leftover argv elements that aren't options. We thus have a
lot of code like this all over the place:

    argc = parse_options(..., 0);
    if (argc)
        usage();

With this flag it becomes:

    parse_options(..., PARSE_OPT_ERROR_AT_NON_OPTION);

Except that we'll also emit a helpful error() about having encountered
unknown non-options.

Change all the callers that were doing this pattern themselves and
passing flags of "0", and that did not emit any custom error()
messages of their own.

Internally to parse_options_step() this is implemented as the
PARSE_OPT_NON_OPTION result. We just need to understand this flag in
parse_options() itself and not return silently in this case.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>